==> src/1.1 Data Gathering.ipynb <==
import pandas as pd
import psycopg2
import pandas.io.sql as sqlio

import config as cfg

host= cfg.db['host']
port = cfg.db['port']
user = cfg.db['user']
pwd = cfg.db['pwd']
database= cfg.db['database']


con = psycopg2.connect(database=database, user=user, password=pwd, host=host, port=port)

print("Database opened successfully")

sql = '''
SELECT * from flights 
where fl_date >= '2018-12-27' and fl_date <= '2019-01-12';
'''

# data = sqlio.read_sql_query(sql, con)

data.to_csv('flights.csv', index=False)

data_samp = data.sample(n = 5000, random_state=0).reset_index()

data_samp.to_csv('flights_samp.csv', index=False)
