==> src/3.1-modelling-linear_regression.ipynb <==
## Linear Regression 

This model was used for a baseline and helped us in the feature selection process. This one is trained with the final training set.from sklearn.ensemble import RandomForestRegressor

from sklearn.linear_model import LinearRegression

from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error

import pandas as pd
import numpy as np

import pickle

# get training data
df = pd.read_csv('../data/encoded_training_data_V7.csv')

# test train fit
y = df.arr_delay.to_numpy()
X = df.drop(columns=['arr_delay']).to_numpy()
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=88)

# fit model
lr = LinearRegression()
lr.fit(X_train, y_train)

# Save pickle file
model = lr
filename = '../model/linear_regression.pkl'
pickle.dump(model, open(filename, 'wb'))

y_pred = lr.predict(X_test)

r2 = r2_score(y_test, y_pred)
MSE = mean_squared_error(y_test,y_pred) 
RMSE = mean_squared_error(y_test,y_pred,squared=False)
MAE = mean_absolute_error(y_test,y_pred)

print(f' r2: {r2}\n MSE: {MSE}\n MAE: {MAE}\n')

# kfold
lm = LinearRegression()
folds = KFold(n_splits = 5, shuffle = True, random_state = 100)
scores = cross_val_score(lm, X_train, y_train, scoring='r2', cv=folds)
scores   

# or MSE
scores = cross_val_score(lm, X_train, y_train, scoring='neg_mean_squared_error', cv=5)
scores

# MAE
scores = cross_val_score(lm, X_train, y_train, scoring='neg_median_absolute_error', cv=5)
scores
