==> src/3.3-modelling-svm.ipynb <==
### SVM Regressor

Note this has been running in various iterations and paired down versions on google colab paid plan, and has timed out ever single time. Will keep trying, but not sure it will complete on time for the submission 

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import make_scorer
from sklearn import svm
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
import pickle
from google.colab import drive
drive.mount('/content/drive')

# df = pd.read_csv("drive/MyDrive/Colab Notebooks/encoded_training_data_v6.csv",low_memory=False, error_bad_lines=False )
df = pd.read_csv("drive/MyDrive/lhl-midterm-data/encoded_training_data_v6.csv" )
df.shape

y = df.arr_delay.to_numpy()
X = df.drop(columns=['arr_delay']).to_numpy()
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=88)

X_train.shape


scorer = make_scorer(mean_squared_error, greater_is_better=False)
# svr_gs = GridSearchCV(SVR(epsilon = 0.01), parameters, cv = K, scoring=scorer)

param_grid = {'C': [1], 'gamma': [0.1],'kernel': ['rbf']}

regr = svm.SVR(epsilon = 0.01) 
grid_search = GridSearchCV(regr, param_grid, cv=3, n_jobs = -1, verbose = 2, scoring=scorer)
grid_search.fit(X_train, y_train)

# Save pickle file
filename = 'svm.pk2'
pickle.dump(grid_search, open(filename, 'wb'))

best_result = grid_search.best_estimator_
print(best_result)

r2 = r2_score(y_test, y_pred)
MSE = mean_squared_error(y_test,y_pred) 
RMSE = mean_squared_error(y_test,y_pred,squared=False)
MAE = mean_absolute_error(y_test,y_pred)

print(f' r2: {r2}\n MSE: {MSE}\n MAE: {MAE}\n')
